+++ scripts/Export-EvergreenAppToFile.ps1
@@ -1,7 +1,7 @@
 <#
     .SYNOPSIS
     Export application details gathered by Get-EvergreenApp to an external JSON file.
-    Reads any existing JSON for that application, adds the new version content, sorts for unique versions, 
+    Reads any existing JSON for that application, adds the new version content, sorts for unique versions,
     then outputs the new content back to the target JSON file.
 #>
 [CmdletBinding(SupportsShouldProcess = $False)]
@@ -76,7 +76,7 @@ process {
             # Export the data to file
             $App | Sort-Object -Property @{ Expression = { [System.Version]$_.Version }; Descending = $false } | `
                 ConvertTo-Json | `
-                Out-File -Path $FilePath -Encoding "Utf8" -NoNewline -Verbose:$VerbosePreference
+                Out-File -FilePath $FilePath -Encoding "Utf8" -NoNewline -Verbose:$VerbosePreference
         }
     }
 }